Add SitesGridItem to DevDocs for easier testing different cases
